requirements.txt
wagtail
wagtailcommonblocks
wagtailmedia
wagtailmenus
django_storages
psycopg2-binary
PyVimeo
unipath
boto3
youtube_dl
gunicorn
